Fix most remaining issues with explosive tags.
damageKnockback failed to use log name, so it could show target name even when
target was out of sight.

Damage projector msg ignored damage_message_passive, so it could show a
message even when the effect was passive and target couldn't be seen.

A bomb placed on a wall could damage either side of the wall.

A bomb on a dead target reverted to its original location.

Saving bombs didn't work properly.